moving resolvers package and fix broken test
