+++ api/kogito-api/src/main/java/org/kie/kogito/correlation/CorrelationResolver.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,13 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.kogito.services.event.impl;
 
-import org.kie.kogito.event.cloudevents.CloudEventExtensionConstants;
+package org.kie.kogito.correlation;
 
-public class KogitoReferenceCorrelationResolver extends SimpleAttributeCorrelationResolver {
+public interface CorrelationResolver {
 
-    public KogitoReferenceCorrelationResolver() {
-        super(CloudEventExtensionConstants.PROCESS_REFERENCE_ID);
-    }
+    Correlation resolve(Object data);
+
 }
