+++ api/kogito-services/src/main/java/org/kie/kogito/services/event/correlation/ExpressionAttributeCorrelationResolver.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.kie.kogito.services.event.correlation;
+
+import org.apache.commons.lang3.NotImplementedException;
+import org.kie.kogito.correlation.Correlation;
+import org.kie.kogito.correlation.CorrelationResolver;
+
+/**
+ * Resolve attributes based on expressions
+ */
+public class ExpressionAttributeCorrelationResolver implements CorrelationResolver {
+
+    @Override
+    public Correlation resolve(Object data) {
+        //todo
+        throw new NotImplementedException();
+    }
+}
