+++ api/kogito-services/src/main/java/org/kie/kogito/services/event/correlation/SimpleAttributeCorrelationResolver.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,17 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.kogito.services.event.impl;
+package org.kie.kogito.services.event.correlation;
 
 import org.kie.kogito.correlation.Correlation;
-import org.kie.kogito.correlation.CorrelationKeyResolver;
+import org.kie.kogito.correlation.CorrelationResolver;
 import org.kie.kogito.event.cloudevents.CloudEventExtensionConstants;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 
-public class SimpleAttributeCorrelationResolver implements CorrelationKeyResolver {
+public class SimpleAttributeCorrelationResolver implements CorrelationResolver {
 
     private String referenceKey = CloudEventExtensionConstants.PROCESS_REFERENCE_ID; //can be expression
     private ObjectMapper objectMapper = new ObjectMapper().registerModule(new JavaTimeModule());
