+++ kogito-codegen-modules/kogito-codegen-processes/src/main/resources/class-templates/MessageConsumerSpringTemplate.java
@@ -58,7 +58,9 @@ public class $Type$MessageConsumer extends AbstractMessageConsumer<$Type$, $Data
 
     private $Type$ eventToModel(Object event) {
         $Type$ model = new $Type$();
-        model.$SetModelMethodName$(objectMapper.convertValue(event, $DataType$.class));
+        if(event != null) {
+            model.$SetModelMethodName$(objectMapper.convertValue(event, $DataType$.class));
+        }
         return model;
     }
